akbank-bootcamp-final/src/main/java/com/mustafagulac/akbank_final_project/weatherapi/controller/WeatherController.java
package com.mustafagulac.akbank_final_project.weatherapi.controller;

import com.mustafagulac.akbank_final_project.weatherapi.model.WeatherResponse;
import com.mustafagulac.akbank_final_project.weatherapi.service.WeatherService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/weather")
public class WeatherController {
    private final WeatherService weatherService;

    @Autowired
    public WeatherController(WeatherService weatherService) {
        this.weatherService = weatherService;
    }

    @GetMapping("/{city}")
    public ResponseEntity<WeatherResponse> getWeatherForecast(@PathVariable String city) {
        WeatherResponse response = weatherService.getWeatherForecast(city);
        if (response != null) {
            return new ResponseEntity<>(response, HttpStatus.OK);
        } else {
            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
        }
    }
}